made fps not always an int lol

+++ src/ui/hairinterface.cpp
@@ -137,7 +137,7 @@ void HairInterface::updateFPSLabel(int totalNumFrames)
     int updateFrequency = 10;
     if (totalNumFrames % updateFrequency == 1) {
         if (totalNumFrames > 1) {
-            int fps = updateFrequency * 1000.0 / m_clock.elapsed();
+            double fps = updateFrequency * 1000.0 / m_clock.elapsed();
             m_ui->fpsLabel->setText(QString::number(fps, 'f', 1) + " FPS");
         }
         m_clock.restart();
